Review of PSH in China (Project dataset and geosptial distribution)
Author: Jingcai Cai 蔡竞蔡

The dataset of PSH projects in China, updated as of April 2024.

Additional information and description shown in sheet 2.

All information are collected by using open-source on Internet. If you want to cite this dataset, please retain the copyright of Jingcai Cai.

This dataset has conduct a map, which also released by me. And the map will also be upload in ArcGIS online soon.

If you want to contact author, could sent e-mal to me to get the dataset and have deeper communication with him.
A map of PSH geospatial distribution in China, updated as of April 2024.

All information are collected by using open-source on Internet. If you want to cite this map, please retain the copyright of Jingcai Cai.

This map will also be upload in ArcGIS online soon, also the dataset will be released soon.

If you want to contact author, could sent e-mal to me to get the dataset and have deeper communication with him.